+++ packages/nimbus/src/components/pagination/index.ts
@@ -1,8 +1,2 @@
-export { Pagination } from "./pagination";
-export type {
-  PaginationProps,
-  PaginationState,
-  UsePaginationProps,
-  UsePaginationReturn,
-} from "./pagination.types";
-export { usePagination } from "./hooks/use-pagination";
+export * from "./pagination";
+export type * from "./pagination.types";
